% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityControl.R
\name{CalMTpercent}
\alias{CalMTpercent}
\title{Calculate the proportion of transcripts mapping to mitochondrial genes}
\usage{
CalMTpercent(SeuratObj, by = "use_internal_data")
}
\arguments{
\item{SeuratObj}{Seurat object}

\item{by}{use_internal_data AnnotationHub biomaRt}
}
\value{

}
\description{
Calculate the proportion of transcripts mapping to mitochondrial genes
}
\examples{
\dontrun{
SeuratObj <- CalMTpercent(SeuratObj, by = 'use_internal_data')
SeuratObj <- CalMTpercent(SeuratObj, by = 'AnnotationHub')
SeuratObj <- CalMTpercent(SeuratObj, by = 'biomaRt')
}

}
