% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_in_shiny.R
\name{check_pkgs}
\alias{check_pkgs}
\title{Check packages for shiny app}
\usage{
check_pkgs(pkgs)
}
\arguments{
\item{pkgs}{packages that need to check for shiny app}
}
\description{
Check packages for shiny app
}
