% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{circleplot}
\alias{circleplot}
\title{circle plot of clusters}
\usage{
circleplot(
  SeuratObj,
  by = "GO",
  pvaluecutoff = 0.01,
  pathwayIDs = NULL,
  link_threshold = 10
)
}
\arguments{
\item{SeuratObj}{Seurat object}

\item{by}{which GSEA result to use, one of "GO", "KEGG", "Reactome", "MSigDb", "WikiPathways", "DO", "NCG", "DGN"}

\item{pvaluecutoff}{calculate link width with pathways lower than pvaluecutoff}

\item{pathwayIDs}{pathway IDs to calculate link width}

\item{link_threshold}{only show links whose intersection number bigger than this threshold}
}
\description{
Link width shows number of intersection of pathways between clusters, link color is
in concordance with cluster which has higher enrichment score on intersection pathways.
}
\examples{
\dontrun{
circleplot(SeuratObj, by = "GO", pvaluecutoff = 0.01, pathwayIDs = NULL)
}


}
