% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicModel_utils.R
\name{cluster_topic_hmp}
\alias{cluster_topic_hmp}
\title{Heatmap showing probability between topics and clusters}
\usage{
cluster_topic_hmp(ldaOut, cluster_rows = T, cluster_cols = T)
}
\arguments{
\item{ldaOut}{topic modeling result}

\item{cluster_rows}{boolean values determining if rows should be clustered.}

\item{cluster_cols}{boolean values determining if columns should be clustered.}
}
\description{
Heatmap showing probability between topics and clusters
}
\examples{
\dontrun{
ldaOut <- SeuratObj@misc$ldaOut
cluster_topic_hmp(ldaOut)
cluster_topic_hmp(ldaOut,  cluster_cols = F)
}

}
