% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{clustercorplot}
\alias{clustercorplot}
\alias{clustercorplot_jaccard}
\title{circle plot of clusters}
\usage{
clustercorplot(
  SeuratObj,
  by = "GO",
  pathwayIDs = NULL,
  color.use = NULL,
  weight.scale = TRUE,
  label.edge = FALSE,
  edge.curved = 0.2,
  shape = "circle",
  layout = igraph::in_circle(),
  margin = 0.1,
  vertex.size.cex = 1,
  link_threshold = 0.5,
  vertex.label.cex = 1.5,
  vertex.label.color = "black",
  arrow.width = 1,
  arrow.size = 0.2,
  edge.label.color = "black",
  edge.label.cex = 0.5,
  edge.max.width = 8,
  vertex.label.dist = 2
)

clustercorplot_jaccard(
  SeuratObj,
  by = "GO",
  pathwayIDs = NULL,
  color.use = NULL,
  weight.scale = TRUE,
  label.edge = FALSE,
  edge.curved = 0.2,
  shape = "circle",
  layout = in_circle(),
  margin = 0.1,
  vertex.size.cex = 1,
  link_threshold = 0.4,
  vertex.label.cex = 1.5,
  vertex.label.color = "black",
  arrow.width = 1,
  arrow.size = 0.2,
  edge.label.color = "black",
  edge.label.cex = 0.5,
  edge.max.width = 8,
  vertex.label.dist = 2
)
}
\arguments{
\item{SeuratObj}{Seurat object}

\item{by}{which GSEA result to use for calculation}

\item{pathwayIDs}{IDs of pathways to use for calculation}

\item{color.use}{used to color nodes, can be a named vector}

\item{weight.scale}{scale the width or not}

\item{label.edge}{label edges or not}

\item{edge.curved}{The degree of edge bending}

\item{shape}{shape of nodes, 'circle' by default}

\item{layout}{network layout, circle by default}

\item{margin}{margin of plot}

\item{vertex.size.cex}{node size}

\item{link_threshold}{only show links whose correlation/Jaccard-index bigger than this threshold}

\item{vertex.label.cex}{size of node label}

\item{vertex.label.color}{color of label of nodes, 'black' by default}

\item{arrow.width}{width of arrow}

\item{arrow.size}{size of arrow}

\item{edge.label.color}{color of label of edge, 'black' by default}

\item{edge.label.cex}{size of edge label}

\item{edge.max.width}{width of edge}

\item{vertex.label.dist}{distance between label and nodes}
}
\value{

}
\description{
Helps to infer the relationship between clusters. Link width shows Pearson correlation or
Jaccard coefficient between clusters, calculated with GSEA result. Node size indicates cell number of each cluster.
If \code{by = "GO"} and \code{pathwayIDs = NULL}, only GO terms of level 5-6 are used for calculation.
}
\examples{
\dontrun{
# show Pearson correlation between clusters
clustercorplot(SeuratObj, by = "GO")
}


\dontrun{
# show Jaccard coefficient between clusters
clustercorplot_jaccard(SeuratObj, by = "GO")
}


}
