% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicModel_predict.R
\name{cosine_heatmap2}
\alias{cosine_heatmap2}
\title{Heatmap showing cosine similarity or pearson correlation between reference data and query data}
\usage{
cosine_heatmap2(
  ldaOut,
  dist,
  a_prefix = "reference_",
  b_prefix = "query_",
  method = c("cosine", "pearson"),
  fontsize = 7
)
}
\arguments{
\item{ldaOut}{reference LDA model}

\item{dist}{transferred model of query data}

\item{a_prefix}{prefix of label of reference data}

\item{b_prefix}{prefix of label of query data}

\item{method}{to calculate cosine similarity or pearson correlation between clusters}

\item{fontsize}{font size}
}
\description{
Heatmap showing cosine similarity or pearson correlation between reference data and query data
}
\examples{
\dontrun{
dist <- query_SeuratObj@misc$dist
cosine_heatmap2(ldaOut=ref_ldaOUt, dist=dist, a_prefix="reference_",
                b_prefix="query_", method="cosine")
}


}
