% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicModel_predict.R
\name{cosine_hive}
\alias{cosine_hive}
\title{Hive diagram showing cosine similarity between reference data and query data}
\usage{
cosine_hive(
  ldaOut,
  dist,
  a_prefix = "reference_",
  b_prefix = "query_",
  node_size = 3,
  node_text_size = 2
)
}
\arguments{
\item{ldaOut}{reference LDA model}

\item{dist}{transferred model of query data}

\item{a_prefix}{prefix of label of reference data}

\item{b_prefix}{prefix of label of query data}

\item{node_size}{node size}

\item{node_text_size}{node text size}
}
\description{
Hive diagram showing cosine similarity between reference data and query data,
calculated with reference LDA model and transferred model.
}
\examples{
\dontrun{
dist <- query_SeuratObj@misc$dist
cosine_hive(ldaOut=ref_ldaOUt, dist=dist, a_prefix="reference_", b_prefix="query_")
}

}
