% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicModel_predict.R
\name{cosine_network_cluster2}
\alias{cosine_network_cluster2}
\title{network showing cosine similarity between reference data and query data}
\usage{
cosine_network_cluster2(
  ldaOut,
  dist,
  a_prefix = "reference_",
  b_prefix = "query_",
  layout = "circle",
  cos_sim_thresh = 0.2,
  SEED = 123,
  radius = 0.12,
  width_range = c(0.1, 0.8),
  text_size = 3
)
}
\arguments{
\item{ldaOut}{reference LDA model}

\item{dist}{transferred model of query data}

\item{a_prefix}{prefix of label of reference data}

\item{b_prefix}{prefix of label of query data}

\item{layout}{layout of network}

\item{cos_sim_thresh}{only links with cosine similarity over threshold will be drawn}

\item{SEED}{seed}

\item{radius}{pie size}

\item{width_range}{edge width}

\item{text_size}{text size}
}
\description{
Network showing cosine similarity between reference data and query data,
calculated with reference LDA model and transferred model.
}
\examples{
\dontrun{
dist <- query_SeuratObj@misc$dist
cosine_network_cluster2(ldaOut=ref_ldaOUt, dist=dist, a_prefix="reference_", b_prefix="query_",
cosine_network_cluster2(ldaOut=ref_ldaOUt, dist=dist, a_prefix="reference_", b_prefix="query_",
}

}
