% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicModel_utils.R
\name{cosineheatmap}
\alias{cosineheatmap}
\title{Heatmap showing cosine similarity between topics}
\usage{
cosineheatmap(ldaOut)
}
\arguments{
\item{ldaOut}{topic modeling result}
}
\description{
calculated by topic~term probability matrix
}
\examples{
\dontrun{
ldaOut <- SeuratObj@misc$ldaOut
cosineheatmap(ldaOut)
}

}
