% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicModel_utils.R
\name{cosinehmp_cluster}
\alias{cosinehmp_cluster}
\title{Heatmap showing cosine similarity between clusters}
\usage{
cosinehmp_cluster(ldaOut)
}
\arguments{
\item{ldaOut}{topic modeling result}
}
\description{
calculated by topic~cluster probability matrix
}
\examples{
\dontrun{
ldaOut <- SeuratObj@misc$ldaOut
cosinehmp_cluster(ldaOut)
}

}
