% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityControl.R
\name{DetectGeneIDtype}
\alias{DetectGeneIDtype}
\title{Detect the gene ID type of Seurat object}
\usage{
DetectGeneIDtype(SeuratObj)
}
\arguments{
\item{SeuratObj}{Seurat object}
}
\value{

}
\description{
Detect the gene ID type of Seurat object
}
\examples{
\dontrun{
SeuratObj <- DetectGeneIDtype(SeuratObj)
GeneIDtype <- slot(object = SeuratObj, name = 'misc')[["featureData"]][['GeneIDtype']]
}

}
