% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{edge_bundling_GSEA}
\alias{edge_bundling_GSEA}
\title{Hierarchical edge bundling plots helps visualizing correlation or similarity between clusters}
\usage{
edge_bundling_GSEA(
  SeuratObj,
  by = "GO",
  link_threshold = 0.6,
  link_width = 0.9,
  method = "cosine similarity",
  node.by = "cluster",
  group.by = "cellType"
)
}
\arguments{
\item{SeuratObj}{Seurat object}

\item{by}{GO KEGG Reactome MSigDb WikiPathways DO NCG DGN.}

\item{link_threshold}{only show links of similarity or correlation above threshold}

\item{link_width}{width of links}

\item{method}{one of \code{c('cosine similarity', 'pearson', 'spearman')}}

\item{node.by}{draw each node as \code{node.by}, column of \code{SeuratObj@meta.data}}

\item{group.by}{color each node by \code{group.by}, column of \code{SeuratObj@meta.data}}
}
\value{

}
\description{
Hierarchical edge bundling plots helps visualizing correlation or similarity between clusters
}
\examples{
edge_bundling_GSEA(SeuratObj, link_threshold=0.6, method='cosine similarity', node.by='cluster', group.by='cellType')
edge_bundling_GSEA(SeuratObj, link_threshold=0.6, method='pearson', node.by='cluster', group.by='cellType')

}
