% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{emapplot2}
\alias{emapplot2}
\title{Network of pathway
Code for plotting is a modified version of \code{enrichplot::emapplot()}.}
\usage{
emapplot2(
  SeuratObj,
  cluster = NULL,
  by = "GO",
  pathwayIDs = NULL,
  showCategory = 20,
  color = "p.adjust",
  layout = "kk",
  node_label_cex = 1,
  node_size_cex = 1,
  cex_line = 1
)
}
\arguments{
\item{SeuratObj}{Object of class "Seurat"}

\item{cluster}{GSEA result of which cluster to show}

\item{by}{which GSEA result to show}

\item{pathwayIDs}{IDs of pathways to show}

\item{showCategory}{number of top pathways of each cluster}

\item{color}{which column to color the nodes, "p.adjust" by default}

\item{layout}{network layout}

\item{node_label_cex}{size of node label}

\item{node_size_cex}{size of node}

\item{cex_line}{width of edge}
}
\value{

}
\description{
Network of pathway
Code for plotting is a modified version of \code{enrichplot::emapplot()}.
}
\examples{
emapplot2(SeuratObj)
emapplot2(SeuratObj, cluster = unique(SeuratObj@misc$GSEAresult_GO$cluster)[2])

}
