% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{emapplotPie}
\alias{emapplotPie}
\title{Network of pathway with node pie}
\usage{
emapplotPie(
  SeuratObj,
  by = "GO",
  pathwayIDs = NULL,
  showCategory = 5,
  color = "p.adjust",
  layout = "kk",
  node_label_cex = 1,
  node_size_cex = 1,
  pie = "-log10FDR",
  cex_line = 1
)
}
\arguments{
\item{SeuratObj}{Object of class "Seurat"}

\item{by}{which GSEA result to show}

\item{pathwayIDs}{IDs of pathways to show}

\item{showCategory}{number of top pathways of each cluster}

\item{color}{which column to color the nodes}

\item{layout}{network layout}

\item{node_label_cex}{size of node label}

\item{node_size_cex}{size of node}

\item{pie}{which to draw pie chart, "-log10FDR" or "count"(number of genes)}

\item{cex_line}{width of edge}
}
\description{
links show jaccard similarity coefficient (overlap of genes) between pathways.
Pie charts show GSEA score of the pathway in clusters. Code for plotting is a modified version of \code{enrichplot::emapplot()}.
If \code{by = "GO"} and \code{pathwayIDs = NULL}, only GO terms of level 5-6 are used for calculation.
}
\examples{
\dontrun{
emapplotPie(SeuratObj)
emapplotPie(SeuratObj, by = "KEGG")
}

}
