% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{embeddedplot}
\alias{embeddedplot}
\title{show embedded histogram or pie chart on UMAP/TSNE plot}
\usage{
embeddedplot(
  SeuratObj,
  by = "GO",
  pathwayIDs = NULL,
  topaths = 1,
  reduction = "umap",
  type = "pie",
  pie.size.cex = 1
)
}
\arguments{
\item{SeuratObj}{Object of class "Seurat"}

\item{by}{which GSEA result to show}

\item{pathwayIDs}{IDs of pathways to show, if provided, parameter \code{topaths} become invalid.}

\item{topaths}{number of top pathways of each cluster to show}

\item{reduction}{"umap", "tsne", "pca"}

\item{type}{"hist", "pie", type of embedded plots}

\item{pie.size.cex}{size of pie chart}
}
\description{
embedded histogram or pie chart show pathway GSEA score on each clusters.
If \code{by = "GO"} and \code{pathwayIDs = NULL}, only GO terms of level 5-6 are used for calculation.
}
\examples{
\dontrun{
embeddedplot(SeuratObj, type = "pie")
embeddedplot(SeuratObj, topaths = 3, reduction = "tsne", type = "hist")
}


}
