% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_kegg_code}
\alias{get_kegg_code}
\title{Get scientific name abbreviate(kegg_code) used in the 'organism' parameter of gseKEGG function}
\usage{
get_kegg_code(species)
}
\arguments{
\item{species}{}
}
\value{

}
\description{
Get scientific name abbreviate(kegg_code) used in the 'organism' parameter of gseKEGG function
}
