% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getEnsDb_AH}
\alias{getEnsDb_AH}
\title{Get EnsDb of a species from package AnnotationHub}
\usage{
getEnsDb_AH(species)
}
\arguments{
\item{species}{}
}
\value{

}
\description{
Get EnsDb of a species from package AnnotationHub
}
