% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getOrgDb_AH}
\alias{getOrgDb_AH}
\title{Get OrgDb of a species from package AnnotationHub}
\usage{
getOrgDb_AH(species)
}
\arguments{
\item{species}{}
}
\value{

}
\description{
Get OrgDb of a species from package AnnotationHub
}
