% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GO2level}
\alias{GO2level}
\title{GO terms and their corresponding level or ontology.}
\format{
A data frame with 137699 rows and 3 variables:
\describe{
\item{GO}{GO IDs}
\item{ont}{ontology}
\item{level}{corresponding level}
}
}
\source{
level information were obtained from \code{getGOLevel()} of \code{clusterProfiler} package.
}
\usage{
GO2level
}
\description{
A dataset containing all GO terms and their corresponding level or ontology.
}
\keyword{datasets}
