% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{GOboxplot}
\alias{GOboxplot}
\title{boxplot showing enrichment score of child or parent GOs of specific GO}
\usage{
GOboxplot(SeuratObj, goid = NULL, type = "child", pointsize = 1, flip = FALSE)
}
\arguments{
\item{SeuratObj}{Object of class "Seurat"}

\item{goid}{specific GO ID}

\item{type}{"child","parent", show child or parent GOs of specific GO}

\item{pointsize}{size of point}

\item{flip}{whether to flip the coordinates}
}
\value{

}
\description{
boxplot showing enrichment score of child or parent GOs of specific GO
}
\examples{
GOboxplot(SeuratObj, goid = "GO:0002576", type = "child")
GOboxplot(SeuratObj, goid = "GO:0002576", type = "parent", flip = T)

}
