% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{GOcircleplot}
\alias{GOcircleplot}
\title{show pathways and genes in chord diagram}
\usage{
GOcircleplot(SeuratObj, genes = NULL)
}
\arguments{
\item{SeuratObj}{Seurat object}

\item{genes}{genes to draw}
}
\description{
show pathways and genes in chord diagram
}
\examples{
\dontrun{
GOcircleplot(SeuratObj, genes=NULL)
}

}
