% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HCL_ldaOut}
\alias{HCL_ldaOut}
\title{built-in reference topic model derived from HCL.}
\format{
A LDA_VEM topic model with 63 topics.
}
\source{
single-cell data comes from \url{https://figshare.com/articles/dataset/HCL_DGE_Data/7235471}
}
\usage{
HCL_ldaOut
}
\description{
built-in reference topic model derived from human cell landscape (HCL).
citation: Han, X., Zhou, Z., Fei, L. et al. Construction of a human cell landscape at single-cell level. Nature 581, 303–309 (2020).
(\url{https://doi.org/10.1038/s41586-020-2157-4})
}
\keyword{datasets}
