% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{hierarchyplot_tree}
\alias{hierarchyplot_tree}
\title{show relationship between clusters and pathways}
\usage{
hierarchyplot_tree(
  SeuratObj,
  by = "GO",
  pathwayIDs = NULL,
  topaths = 5,
  cluster_cutree_k = NULL,
  pathway_cutree_k = NULL,
  color.use.cluster = NULL,
  color.use.pathway = NULL,
  weight.scale = TRUE,
  vertex.size.cex = 1,
  vertex.label.cex = 0.9,
  edge.max.width = 3,
  vertex.label.color = "black",
  alpha.edge = 0.6
)
}
\arguments{
\item{SeuratObj}{Object of class "Seurat"}

\item{by}{which GSEA result to show}

\item{pathwayIDs}{IDs of pathways to show}

\item{topaths}{top n pathway of each cluster}

\item{cluster_cutree_k}{the clusters are divided into k, based on the hierarchical clustering (using cutree)}

\item{pathway_cutree_k}{the pathways are divided into k, based on the hierarchical clustering (using cutree)}

\item{color.use.cluster}{used to color the cluster nodes, a named vector or a vector}

\item{color.use.pathway}{used to color the pathway nodes}

\item{weight.scale}{scale the width or not}

\item{vertex.size.cex}{node size}

\item{vertex.label.cex}{size of node label}

\item{edge.max.width}{width of edge}

\item{vertex.label.color}{color of label of nodes, 'black' by default}

\item{alpha.edge}{transparency of edge color}
}
\description{
Hierarchical clustering of clusters and pathways according to GSEA result.
If \code{by = "GO"} and \code{pathwayIDs = NULL}, only GO terms of level 5-6 are used for calculation.
}
\examples{
\dontrun{
hierarchyplot_tree(SeuratObj, by = "GO", topaths = 10, cluster_cutree_k = 5, pathway_cutree_k = 20)
}


}
