% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicModel_utils.R
\name{hist_cluster_topic}
\alias{hist_cluster_topic}
\title{barplot to show probability of assigned topics in clusters}
\usage{
hist_cluster_topic(gammaDF, clusters)
}
\arguments{
\item{gammaDF}{cluster~topic probability data.frame}

\item{clusters}{clusters to show}
}
\value{
ggplot object
}
\description{
barplot to show probability of assigned topics in clusters
}
\examples{
ldaOut <- SeuratObj@misc$ldaOut
gammaDF <- tidytext::tidy(ldaOut, matrix = "gamma")
hist_cluster_topic(gammaDF, clusters=head(Seurat::Idents(SeuratObj)))

}
