% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicModel_utils.R
\name{hist_topic_term}
\alias{hist_topic_term}
\title{barplot to show top terms of every topic}
\usage{
hist_topic_term(betaDF, topics, topn = 5, axis.text.y.size = 5)
}
\arguments{
\item{betaDF}{topic~term probability data.frame}

\item{topics}{topics to show}

\item{topn}{top n terms of topic}

\item{axis.text.y.size}{size of y axis text}
}
\value{
ggplot object
}
\description{
barplot to show top terms of every topic
}
\examples{
\dontrun{
ldaOut <- SeuratObj@misc$ldaOut
betaDF <- tidytext::tidy(ldaOut, matrix = "beta")
pws <- SeuratObj@misc$GSEAresult_GO \%>\% dplyr::select(ID, Description) \%>\% unique \%>\% tibble::deframe()
betaDF \%<>\% dplyr::mutate(descrip=unname(pws[term]))
hist_topic_term(betaDF, topics=1, topn=20, axis.text.y.size=10)
hist_topic_term(betaDF, topics=c(1,2,3,4), topn=20, axis.text.y.size=10)
}


}
