% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ID2Description}
\alias{ID2Description}
\title{Get pathway IDs and their corresponding description}
\usage{
ID2Description(SeuratObj, by = "GO")
}
\arguments{
\item{SeuratObj}{Object of class "Seurat"}

\item{by}{"GO", "KEGG", "Reactome", "MSigDb", "WikiPathways", "DO", "NCG", "DGN"}
}
\description{
Get pathway IDs and their corresponding description
}
\examples{
\dontrun{
ID2Description(SeuratObj, by = "GO")
ID2Description(SeuratObj, by = "KEGG")
}

}
