% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicModel_utils.R
\name{lda.rep}
\alias{lda.rep}
\title{calculate metrics to find the best number of topics}
\usage{
lda.rep(dtm, max.k = 5, method = "VEM", SEED = 1234)
}
\arguments{
\item{dtm}{Object of class "DocumentTermMatrix"}

\item{max.k}{max number of k}

\item{method}{The method to be used for fitting; currently method = "VEM" or method= "Gibbs" are supported.}

\item{SEED}{seed}
}
\value{
list of 4 metrics: perplexity, loglikelihood, alpha, entropy, cosine similarity
}
\description{
calculate metrics to find the best number of topics
}
\examples{
\dontrun{
ldas <- lda.rep(dtm=dtm, max.k=10, method='VEM', SEED=123)
plotindex(ldas)
}


}
