% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MCA_ldaOut}
\alias{MCA_ldaOut}
\title{built-in reference topic model derived from MCA.}
\format{
A LDA_VEM topic model with 52 topics.
}
\source{
single-cell data comes from \url{https://figshare.com/articles/dataset/HCL_DGE_Data/7235471}
}
\usage{
MCA_ldaOut
}
\description{
built-in reference topic model derived from Mouse Cell Atlas (MCA).
citation: \url{https://doi.org/10.1016/j.cell.2018.02.001}.
}
\keyword{datasets}
