% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MTGENE}
\alias{MTGENE}
\title{Mitochondrial genes of 3 species.}
\format{
A list containing 3 elements:
\describe{
\item{Homo sapiens}{mitochondrial genes of several gene ID types}
\item{Mus musculus}{mitochondrial genes of several gene ID types}
\item{Arabidopsis thaliana}{mitochondrial genes of several gene ID types}
}
}
\source{
mitochondrial genes information derived from bioMart and NCBI.
}
\usage{
MTGENE
}
\description{
A list containing mitochondrial genes of 3 species, Homo sapiens, Mus musculus and Arabidopsis thaliana.
}
\keyword{datasets}
