% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicModel_predict.R
\name{oversample}
\alias{oversample}
\title{Oversample the reference data}
\usage{
oversample(
  reference_SeuratObj,
  number_clusters = NULL,
  group_by = "cellType",
  cluster_by = "seurat_clusters"
)
}
\arguments{
\item{reference_SeuratObj}{reference data}

\item{number_clusters}{goal number of clusters that you need to oversample to}

\item{group_by}{the column of \code{reference_SeuratObj@meta.data} that indicates cell type}

\item{cluster_by}{the column of \code{reference_SeuratObj@meta.data} that indicates cluster}
}
\value{
a Seurat object with oversampled expression matrix
}
\description{
In order to avoid class imbalance while training a classifier, we oversampled the training set.
In other words, we randomly selected cells from clusters of a particular cell type to form new clusters,
reaching the situation that each cell type owns the same number of clusters
}
