% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicModel_predict.R
\name{oversample_ref}
\alias{oversample_ref}
\title{Oversample the reference data}
\usage{
oversample_ref(
  reference_SeuratObj,
  number_clusters = NULL,
  group_by = "cellType",
  cluster_by = "seurat_clusters",
  species = "Homo sapiens",
  by = "GO",
  k = NULL,
  method = "VEM"
)
}
\arguments{
\item{reference_SeuratObj}{reference data}

\item{number_clusters}{goal number of clusters that you need to oversample to}

\item{group_by}{the column of \code{reference_SeuratObj@meta.data} that indicates cell type}

\item{cluster_by}{the column of \code{reference_SeuratObj@meta.data} that indicates cluster}

\item{species}{species of the reference data}

\item{by}{database used to perform GSEA. GO KEGG Reactome MSigDb WikiPathways DO NCG DGN.}

\item{k}{number of topics.}

\item{method}{method used for fitting a LDA model; currently "VEM" or "Gibbs" are supported.}
}
\value{
a Seurat object with oversampled expression matrix and topic-model result.
}
\description{
Oversample the reference data and perform GSEA, topic modelling.
In order to avoid class imbalance while training a classifier, we oversampled the training set.
In other words, we randomly selected cells from clusters of a particular cell type to form new clusters,
reaching the situation that each cell type owns the same number of clusters.
}
\examples{
\dontrun{
reference_SeuratObj <- oversample_ref(reference_SeuratObj, number_clusters = 10,
group_by = 'cellType', cluster_by = 'seurat_clusters',
species = "Homo sapiens", by = 'GO', k = NULL, method = "VEM")
}


}
