% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{pathwayScatterplot}
\alias{pathwayScatterplot}
\title{Scatter plot showing pathway enrichment score}
\usage{
pathwayScatterplot(
  SeuratObj,
  by = "GO",
  pathwayID = NULL,
  reduction = "umap",
  colour = "OrRd",
  pointsize = 1,
  label = TRUE,
  label.size = 4
)
}
\arguments{
\item{SeuratObj}{Object of class "Seurat"}

\item{by}{which GSEA result to show}

\item{pathwayID}{ID of pathway to show}

\item{reduction}{"umap", "tsne", "pca"}

\item{colour}{see \code{RColorBrewer::brewer.pal.info}}

\item{pointsize}{size of point}

\item{label}{label cluster or not}

\item{label.size}{size of label}
}
\description{
Scatter plot showing pathway enrichment score
}
\examples{
\dontrun{
pathwayScatterplot(SeuratObj, by = "GO", pathwayID = "GO:0002576")
}

}
