% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicModel_utils.R
\name{plot_sankey}
\alias{plot_sankey}
\title{Sankey diagram showing best assigned topic of each cluster}
\usage{
plot_sankey(gammaDF, topn = 1, plotHeight = 600)
}
\arguments{
\item{gammaDF}{cluster~topic probability data.frame}

\item{topn}{top n terms of each topic}

\item{plotHeight}{height of plot}
}
\value{

}
\description{
Sankey diagram showing best assigned topic of each cluster
}
\examples{
ldaOut <- SeuratObj@misc$ldaOut
gammaDF <- tidytext::tidy(ldaOut, matrix = "gamma")
plot_sankey(gammaDF, topn=1)
plot_sankey(gammaDF, topn=2)

}
