% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicModel_utils.R
\name{plotindex}
\alias{plotindex}
\title{plot metrics to find the best number of topics}
\usage{
plotindex(ldas)
}
\arguments{
\item{ldas}{list of 4 metrics(perplexity, loglikelihood, alpha, entropy, cosine similarity), result of \code{lda.rep()}.}
}
\value{
ggplot object
}
\description{
plot metrics to find the best number of topics
}
\examples{
\dontrun{
ldas <- lda.rep(dtm=dtm, max.k=10, method='VEM', SEED=123)
plotindex(ldas)
}


}
