% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicModel_utils.R
\name{plotnw1}
\alias{plotnw1}
\title{network showing top terms of topics}
\usage{
plotnw1(betaDF, topn)
}
\arguments{
\item{betaDF}{topic~term probability data.frame}

\item{topn}{number of top terms of each topic}
}
\value{

}
\description{
network showing top terms of topics
}
\examples{
ldaOut <- SeuratObj@misc$ldaOut
betaDF <- tidytext::tidy(ldaOut, matrix = "beta")
plotnw1(betaDF, topn=10)

}
