% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicModel_utils.R
\name{plotnw3}
\alias{plotnw3}
\title{echart4r network showing top terms of topics}
\usage{
plotnw3(betaDF, topn)
}
\arguments{
\item{betaDF}{topic~term probability data.frame}

\item{topn}{top n terms of topic}
}
\description{
echart4r network showing top terms of topics
}
\examples{
ldaOut <- SeuratObj@misc$ldaOut
betaDF <- tidytext::tidy(ldaOut, matrix = "beta")
plotnw3(betaDF, topn=10)

}
