% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicModel_predict.R
\name{predictFun}
\alias{predictFun}
\title{Predict cell types of query data based on reference data}
\usage{
predictFun(
  query_SeuratObj,
  reference_SeuratObj,
  group_by = "cellType",
  cluster_by = "seurat_clusters",
  species = "Homo sapiens",
  by = "GO",
  k = NULL,
  LDAmethod = "VEM"
)
}
\arguments{
\item{query_SeuratObj}{query data}

\item{reference_SeuratObj}{reference data}

\item{group_by}{the column of \code{reference_SeuratObj@meta.data} that indicates cell type}

\item{cluster_by}{the column of \code{reference_SeuratObj@meta.data} that indicates cluster}

\item{species}{species of the reference data}

\item{by}{database used to perform GSEA on reference data. This parameter works when \code{reference_SeuratObj} does not contain topic model.}

\item{k}{number of topics. This parameter works when \code{reference_SeuratObj} does not contain topic model.}

\item{LDAmethod}{method used for fitting a LDA model on reference data; currently "VEM" or "Gibbs" are supported. This parameter works when \code{reference_SeuratObj} does not contain topic model.}
}
\value{
data frame with 2 columns including clusters of query data, and the predicted cell type based on reference data.
}
\description{
Predict cell type of query data based on reference data. We use caret R package to train a svm classifier.
}
\examples{
\dontrun{
predictFun(query_SeuratObj, reference_SeuratObj, group_by = 'cellType', cluster_by = 'seurat_clusters',
species = "Homo sapiens", by = 'GO', k = NULL, LDAmethod = "VEM")
}


}
