% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{readData}
\alias{readData}
\title{Load data from different sources}
\usage{
readData(
  data,
  type = c("10X", "10X_h5", "expMatrix", "SingleCellExperiment", "Seurat",
    "CellDataSet", "loom"),
  meta.data = NULL,
  species
)
}
\arguments{
\item{data}{Can be a directory like "/outs/filtered_gene_bc_matrices/" provided by 10X.
Or an expression matrix of single-cell RNA-seq (matrix or dgCMatrix).
Or an object from R/python packages such as scater, scran, Seurat, monocle.}

\item{type}{10X, 10X_h5, expMatrix, SingleCellExperiment, Seurat, CellDataSet, loom}

\item{meta.data}{Additional cell-level metadata to add to the Seurat object. Should be a data frame where the rows are cell names and the columns are additional metadata fields.}

\item{species}{Species name, such as Homo sapiens or Mus musculus.}
}
\value{
Seurat object
}
\description{
Load data from different sources
}
\examples{
\dontrun{
SeuratObj <- CellFunMap::readData(data = "~/outs/filtered_gene_bc_matrices/GRCh38/",
                                  type = "10X", species = "Homo sapiens")
SeuratObj <- CellFunMap::readData(data = counts, type = 'expMatrix', species = "Homo sapiens")
}


}
