% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunGSEA.R
\name{RunGSEA}
\alias{RunGSEA}
\title{Perform Gene Set Enrichment Analysis (GSEA) on Seurat object}
\usage{
RunGSEA(
  SeuratObj,
  by = "GO",
  TERM2GENE = NULL,
  minpct = 0,
  pvalueCutoff = 1,
  category = NULL,
  subcategory = NULL
)
}
\arguments{
\item{SeuratObj}{Seurat object}

\item{by}{GO KEGG Reactome MSigDb WikiPathways DO NCG DGN. May be a character vector. Will be ignored if parameter "TERM2GENE" is not NULL.}

\item{TERM2GENE}{Customized terms. Annotation of TERM TO GENE mapping, a data.frame of 2 column with term and gene. NULL by default. If not NULL, parameter "by" will be ignored.}

\item{minpct}{minimum expression percent in each cluster, parameter used to filter Differential expression genes to run GSEA. 0 as default.}

\item{pvalueCutoff}{threshold of p-value in GSEA result}

\item{category}{MSigDB collection abbreviation, such as H, C1, C2, C3, C4, C5, C6, C7. If NULL, all gene sets.}

\item{subcategory}{MSigDB sub-collection abbreviation, such as CGP or BP.}
}
\value{
Seurat object
}
\description{
Perform Gene Set Enrichment Analysis (GSEA) on Seurat object. GSEA is implemented using clusterProfiler package.
}
\examples{
\dontrun{
SeuratObj <- RunGSEA(SeuratObj, by = 'GO')
}

}
