% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunSeurat.R
\name{RunSeurat}
\alias{RunSeurat}
\title{basic single cell analysis with Seurat}
\usage{
RunSeurat(SeuratObj, nPCs = 10, resolution = 0.5)
}
\arguments{
\item{SeuratObj}{Seurat object}

\item{nPCs}{Dimensions of reduction to use as input}

\item{resolution}{Value of the resolution parameter, use a value above (below) 1.0 if you want to obtain a larger (smaller) number of communities.}
}
\value{
Seurat object
}
\description{
basic single cell analysis with Seurat
}
\examples{
\dontrun{
SeuratObj <- RunSeurat(SeuratObj, nPCs = 10, resolution = 0.5)
}



}
