% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicModel_predict.R
\name{sankey_comparison}
\alias{sankey_comparison}
\title{Sankey diagram comparing the reference LDA model and query data}
\usage{
sankey_comparison(
  ldaOut,
  dist,
  a_prefix = "reference_",
  b_prefix = "query_",
  height = 1600,
  width = 800,
  fontSize = 12
)
}
\arguments{
\item{ldaOut}{reference LDA model}

\item{dist}{transferred model of query data}

\item{a_prefix}{prefix of label of reference data}

\item{b_prefix}{prefix of label of query data}

\item{height}{output height of sankey diagram}

\item{width}{output width of sankey diagram}

\item{fontSize}{font size}
}
\value{

}
\description{
After we transfer the reference LDA model to query data (\code{transfer_LDA()}), we can use this function to visualize
the comparison of reference LDA model and query data.
}
\examples{
\dontrun{
dist <- query_SeuratObj@misc$dist
sankey_comparison(ldaOut=ref_ldaOUt, dist=dist, a_prefix="reference_", b_prefix="query_") \%>\% saveNetwork(file = "sankey_comparison.html")
}


}
