% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityControl.R
\name{ScFunQC}
\alias{ScFunQC}
\title{Title
quality control according to 'nCount_RNA', 'nFeature_RNA', 'percent.mt', 'CellsPerGene'}
\usage{
ScFunQC(
  SeuratObj,
  minCountsPerCell = NULL,
  maxCountsPerCell = NULL,
  minFeaturesPerCell = NULL,
  maxFeaturesPerCell = NULL,
  maxPercent.mt = NULL
)
}
\arguments{
\item{SeuratObj}{Seurat object}

\item{minCountsPerCell}{minCountsPerCell}

\item{maxCountsPerCell}{maxCountsPerCell}

\item{minFeaturesPerCell}{minFeaturesPerCell}

\item{maxFeaturesPerCell}{maxFeaturesPerCell}

\item{maxPercent.mt}{maxPercent.mt}
}
\value{

}
\description{
Title
quality control according to 'nCount_RNA', 'nFeature_RNA', 'percent.mt', 'CellsPerGene'
}
\examples{

}
