% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Colors.R
\name{scPalette3}
\alias{scPalette3}
\title{Generate colors from a customed color palette}
\usage{
scPalette3(n)
}
\arguments{
\item{n}{number of colors to generate}
}
\value{

}
\description{
Generate colors from a customed color palette
}
\examples{
scPalette3(8)
pals::pal.bands(scPalette3(20))

}
