% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{simplifyEnrichment}
\alias{simplifyEnrichment}
\title{cluster functional terms into groups by clustering the similarity matrix of the terms}
\usage{
simplifyEnrichment(pathwayid, by = "GO")
}
\arguments{
\item{pathwayid}{a vector of pathway IDs, such as a vector of GO IDs}

\item{by}{GO, KEGG, Reactome, MSigDb}
}
\value{

}
\description{
A new method (binary cut) is proposed to effectively cluster GO terms into groups from the semantic similarity matrix. Summaries of GO terms in each cluster are visualized by word clouds.
https://bioconductor.org/packages/release/bioc/html/simplifyEnrichment.html
https://bioconductor.org/packages/release/bioc/vignettes/simplifyEnrichment/inst/doc/simplifyEnrichment.html
package "simplifyEnrichment" is developed based on R version 4.0
}
\examples{
\dontrun{
toppws <- SeuratObj@misc$GSEAresult_GO \%>\% dplyr::group_by(cluster) \%>\% dplyr::slice_min(order_by = p.adjust, n = 10, with_ties = F) \%>\% dplyr::pull(ID) \%>\% unique
simplifyEnrichment(toppws)
}


}
