% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{simplifyEnrichmentplot}
\alias{simplifyEnrichmentplot}
\title{cluster functional terms into groups by clustering the similarity matrix of the terms}
\usage{
simplifyEnrichmentplot(
  SeuratObj,
  by = "GO",
  pathwayIDs = NULL,
  showCategory = 10,
  GO_ont = "BP"
)
}
\arguments{
\item{SeuratObj}{Seurat object}

\item{by}{GO, KEGG, Reactome, MSigDb}

\item{pathwayIDs}{a vector of pathway IDs, such as a vector of GO IDs}

\item{showCategory}{number of top pathways of each cluster}

\item{GO_ont}{Gene Ontology}
}
\description{
We utilize package simplifyEnrichment to cluster GO terms into groups from the semantic similarity matrix. Summaries of GO terms in each cluster are visualized by word clouds.
https://bioconductor.org/packages/release/bioc/html/simplifyEnrichment.html
https://bioconductor.org/packages/release/bioc/vignettes/simplifyEnrichment/inst/doc/simplifyEnrichment.html
package "simplifyEnrichment" is developed based on R version 4.0
}
\examples{
\dontrun{
simplifyEnrichmentplot(SeuratObj, by = "GO", showCategory = 10)
}


}
