% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{speToOrg}
\alias{speToOrg}
\title{get 'organism' that ReactomePA::gsePathway needs.}
\usage{
speToOrg(species)
}
\arguments{
\item{species}{species name}
}
\description{
change latin name of species to trivial name
}
