% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicModel_utils.R
\name{topicProb}
\alias{topicProb}
\title{Show topic activity on UMAP/TSNE cell map}
\usage{
topicProb(SeuratObj, reduction = "umap", topic = 1, pointSize = 0.1)
}
\arguments{
\item{SeuratObj}{Seurat object}

\item{reduction}{umap or tsne}

\item{topic}{which topic to show}

\item{pointSize}{0.1 default.}
}
\value{
ggplot object
}
\description{
Show topic activity on UMAP/TSNE cell map
}
\examples{
\dontrun{
topicProb(SeuratObj, reduction="umap", topic=3, pointSize=0.1)
}

}
