% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicModel_utils.R
\name{Topterms_Topic}
\alias{Topterms_Topic}
\title{Show top terms of each topic}
\usage{
Topterms_Topic(betaDF, Topic, topn = 10, text.size = 3)
}
\arguments{
\item{betaDF}{topic~term probability data.frame}

\item{Topic}{topic to show}

\item{topn}{top n terms of topic}

\item{text.size}{size of text}
}
\value{
ggplot object
}
\description{
Show top terms of each topic
}
\examples{
\dontrun{
ldaOut <- SeuratObj@misc$ldaOut
betaDF <- tidytext::tidy(ldaOut, matrix = "beta")
pws <- ID2Description(SeuratObj, by = "GO")
betaDF \%<>\% dplyr::mutate(descrip=unname(pws[term]))
Topterms_Topic(betaDF, Topic = 1, topn = 10, text.size = 3)
}

}
