% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicModel_predict.R
\name{transfer_LDA}
\alias{transfer_LDA}
\title{Transfer LDA model from reference data to query data}
\usage{
transfer_LDA(query_SeuratObj, ref_ldaOUt, by = "GO")
}
\arguments{
\item{query_SeuratObj}{query data}

\item{ref_ldaOUt}{reference LDA model}

\item{by}{using which GSEA result of query data to transfer, one of "GO", "KEGG", "Reactome", "MSigDb", "WikiPathways", "DO", "NCG", "DGN"}
}
\value{
a Seurat object. Transferred LDA model of query data is stored in \code{query_SeuratObj@misc$dist}, which
is a list containing clusters-by-topics (θ) and topics-by-pathways (ϕ) matrices.
}
\description{
Transfer LDA model from reference data to query data
}
\examples{
\dontrun{
query_SeuratObj <- transfer_LDA(query_SeuratObj, ref_ldaOUt, by = "GO")
}

}
