% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{txid_gene_MT}
\alias{txid_gene_MT}
\title{Mitochondrial genes of all species.}
\format{
A data frame with 310200 rows and 4 variables:
\describe{
\item{txid}{taxonomy id of species}
\item{ENTREZID}{ENTREZID of genes}
\item{SYMBOL}{symbol of genes}
\item{chromosome}{MT}
}
}
\source{
\url{ftp://ftp.ncbi.nlm.nih.gov/gene/DATA/GENE_INFO/All_Data.gene_info.gz}
}
\usage{
txid_gene_MT
}
\description{
A dataset containing mitochondrial genes of all species in NCBI.
}
\keyword{datasets}
