% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_in_shiny.R
\name{visualize_in_shiny}
\alias{visualize_in_shiny}
\title{visualize the Seurat object in a built-in shiny app}
\usage{
visualize_in_shiny(SeuratObj)
}
\arguments{
\item{SeuratObj}{Seurat object}
}
\description{
visualize the Seurat object in a built-in shiny app
}
\examples{
\dontrun{
if(interactive()) {
    visualize_in_shiny(SeuratObj)
}
}

}
