% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicModel_utils.R
\name{wordcloud_topic}
\alias{wordcloud_topic}
\alias{wordcloud_topic_3D}
\title{word cloud of top terms of topic}
\usage{
wordcloud_topic(betaDF, pws, topic, topn = 20)

wordcloud_topic_3D(betaDF, pws, topic, topn = 20)
}
\arguments{
\item{betaDF}{topic~term probability data.frame}

\item{pws}{named vector, pathway description with ID as names}

\item{topic}{topic to draw}

\item{topn}{top n terms of topic}
}
\description{
word cloud of top terms of topic
}
\examples{
\dontrun{
ldaOut <- SeuratObj@misc$ldaOut
betaDF <- tidytext::tidy(ldaOut, matrix = "beta")
pws <- SeuratObj@misc$GSEAresult_GO \%>\% dplyr::select(ID, Description) \%>\% unique \%>\% tibble::deframe()
wordcloud_topic(betaDF, pws, topic=1, topn=20)
wordcloud_topic_3D(betaDF, pws, topic=1, topn=20)
}

}
