% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicModel_utils.R
\name{cosine_network_cluster}
\alias{cosine_network_cluster}
\title{Network showing cosine similarity between clusters}
\usage{
cosine_network_cluster(
  ldaOut,
  layout = "circle",
  cos_sim_thresh = 0.2,
  SEED = 123,
  radius = 0.12,
  width_range = c(0.1, 0.8),
  text_size = 3
)
}
\arguments{
\item{ldaOut}{topic modeling result}

\item{layout}{network layout}

\item{cos_sim_thresh}{only shows edge of cosine similarity greater than specific threshold}

\item{SEED}{seed}

\item{radius}{node pie radius}

\item{width_range}{range of width of edges}

\item{text_size}{size of labels}
}
\value{
ggplot object
}
\description{
Network of cosine similarity between clusters, calculated by cluster~topic probability matrix. Width of edge shows cosine similarity between clusters, node pie shows topic probability distribution of each cluster.
}
\examples{
\dontrun{
ldaOut <- SeuratObj@misc$ldaOut
cosine_network_cluster(ldaOut, layout="circle", cos_sim_thresh=0.2, SEED=123,
                       radius=0.12, width_range=c(0.1, 0.8))
cosine_network_cluster(ldaOut, layout="fr", cos_sim_thresh=0.2, SEED=123,
                       radius=0.12, width_range=c(0.1, 0.8))
}

}
