% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicModel_utils.R
\name{cosineheatmap}
\alias{cosineheatmap}
\title{Heatmap showing cosine similarity between topics}
\usage{
cosineheatmap(ldaOut)
}
\arguments{
\item{ldaOut}{topic modeling result}
}
\description{
calculated by topic~term probability matrix
}
\examples{
ldaOut <- SeuratObj@misc$ldaOut
# ldaOut <- topicmodels::LDA(dtm, k = 10, method = "VEM", control = list(seed = 1234))
cosineheatmap(ldaOut)

}
