% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityControl.R
\name{DetectGeneIDtype}
\alias{DetectGeneIDtype}
\title{Title
detect what gene ID type that rownames(counts) belongs to}
\usage{
DetectGeneIDtype(SeuratObj)
}
\arguments{
\item{SeuratObj}{Seurat object}
}
\value{

}
\description{
Title
detect what gene ID type that rownames(counts) belongs to
}
\examples{
#' \dontrun{
SeuratObj <- DetectGeneIDtype(SeuratObj)
GeneIDtype <- slot(object = SeuratObj, name = 'misc')[["featureData"]][['GeneIDtype']]
}

}
