% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{GOcircleplot}
\alias{GOcircleplot}
\title{show pathways and genes in chord diagram}
\usage{
GOcircleplot(SeuratObj, Orgdb = org.Hs.eg.db, genes = NULL)
}
\arguments{
\item{SeuratObj}{Seurat object}

\item{Orgdb}{Orgdb of specific species}

\item{genes}{genes to draw}
}
\value{

}
\description{
show pathways and genes in chord diagram
}
\examples{
GOcircleplot(SeuratObj, Orgdb=org.Hs.eg.db, genes=NULL)

}
