% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{speToOrg}
\alias{speToOrg}
\title{get 'organism' that ReactomePA::gsePathway needs.
change latin name of species to trivial name}
\usage{
speToOrg(species)
}
\arguments{
\item{species}{}
}
\value{

}
\description{
get 'organism' that ReactomePA::gsePathway needs.
change latin name of species to trivial name
}
