% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicModel_utils.R
\name{topicNW3}
\alias{topicNW3}
\title{echart4r network showing top terms of topics}
\usage{
topicNW3(betaDF, topn, pws)
}
\arguments{
\item{betaDF}{topic~term probability data.frame}

\item{topn}{top n terms of topic}

\item{pws}{named vector, pathway description with ID as names}
}
\description{
echart4r network showing top terms of topics with tooltips showing description of pathway ID
}
\examples{
\dontrun{
ldaOut <- SeuratObj@misc$ldaOut
betaDF <- tidytext::tidy(ldaOut, matrix = "beta")
pws <- ID2Description(SeuratObj, by = "GO")
topicNW3(betaDF, topn=10, pws)
}

}
