% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicModel_utils.R
\name{umap_cluster}
\alias{umap_cluster}
\title{UMAP on cluster-topic probability matrix}
\usage{
umap_cluster(ldaOut, label.size = 3)
}
\arguments{
\item{ldaOut}{topic modeling result}

\item{label.size}{size of labels}
}
\description{
Each point represents clusters. Point color indicates the topic with the highest probability of each cluster.
}
\examples{
\dontrun{
ldaOut <- SeuratObj@misc$ldaOut
umap_cluster(ldaOut)
}


}
