% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicModel_utils.R
\name{umap_cluster}
\alias{umap_cluster}
\title{UMAP on cluster-topic probability matrix
each point represents clusters}
\usage{
umap_cluster(ldaOut)
}
\arguments{
\item{ldaOut}{topic modeling result}
}
\value{

}
\description{
UMAP on cluster-topic probability matrix
each point represents clusters
}
\examples{
ldaOut <- SeuratObj@misc$ldaOut
# ldaOut <- topicmodels::LDA(dtm, k = 10, method = "VEM", control = list(seed = 1234))
umap_cluster(ldaOut)


}
